requirements.txt
# Core dependencies
haystack-ai>=2.0.0
mysql-connector-python>=8.0.0
chromadb>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# MCP dependencies
mcp>=1.0.0

# LLM dependencies
ollama>=0.2.0

# Async support
asyncio-mqtt
aiofiles

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0